added spectral normalization to rest of custom building blocks
